fix: Convert Step object's id to index
The IDs passed by the network were not consistently incremented. They started at 0 and initially appeared to increment as an index counter. There was at least one ID that was equal to the number of steps in that recipe. This caused an Out-of-Bounds error when the Adapter attempted to get the ViewHolder at that ID.